LPS-190477 feature-flag: refactor -web into -api and -impl
