+++ modules/apps/feature-flag/feature-flag-web/src/main/java/com/liferay/feature/flag/web/internal/display/FeatureFlagDisplay.java
@@ -1,9 +1,12 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.feature.flag.web.internal.model;
+package com.liferay.feature.flag.web.internal.display;
+
+import com.liferay.feature.flag.model.FeatureFlag;
+import com.liferay.feature.flag.model.FeatureFlagWrapper;
 
 import java.util.List;
 import java.util.Locale;
