--- modules/apps/feature-flag/feature-flag-web/src/main/java/com/liferay/feature/flag/web/internal/manager/FeatureFlagPreferencesManager.java
@@ -1,67 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.feature.flag.web.internal.manager;
-
-import com.liferay.feature.flag.web.internal.constants.FeatureFlagConstants;
-import com.liferay.portal.kernel.portlet.PortalPreferences;
-import com.liferay.portal.kernel.service.PortalPreferencesLocalService;
-import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.PortletKeys;
-import com.liferay.portal.kernel.util.Validator;
-import com.liferay.portlet.PortalPreferencesWrapper;
-
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
-
-/**
- * @author Drew Brokke
- */
-@Component(service = FeatureFlagPreferencesManager.class)
-public class FeatureFlagPreferencesManager {
-
-	public Boolean isEnabled(long companyId, String key) {
-		if (Validator.isNull(
-				_portalPreferencesLocalService.fetchPortalPreferences(
-					companyId, PortletKeys.PREFS_OWNER_TYPE_COMPANY))) {
-
-			return null;
-		}
-
-		PortalPreferences portalPreferences = _getPortalPreferences(companyId);
-
-		String value = portalPreferences.getValue(_NAMESPACE, key);
-
-		if (value == null) {
-			return null;
-		}
-
-		return GetterUtil.getBoolean(value);
-	}
-
-	public void setEnabled(long companyId, String key, boolean enabled) {
-		PortalPreferences portalPreferences = _getPortalPreferences(companyId);
-
-		portalPreferences.setValue(_NAMESPACE, key, String.valueOf(enabled));
-
-		_portalPreferencesLocalService.updatePreferences(
-			companyId, PortletKeys.PREFS_OWNER_TYPE_COMPANY, portalPreferences);
-	}
-
-	private PortalPreferences _getPortalPreferences(long companyId) {
-		PortalPreferencesWrapper portalPreferencesWrapper =
-			(PortalPreferencesWrapper)
-				_portalPreferencesLocalService.getPreferences(
-					companyId, PortletKeys.PREFS_OWNER_TYPE_COMPANY);
-
-		return portalPreferencesWrapper.getPortalPreferencesImpl();
-	}
-
-	private static final String _NAMESPACE = FeatureFlagConstants.FEATURE_FLAG;
-
-	@Reference
-	private PortalPreferencesLocalService _portalPreferencesLocalService;
-
-}
