+++ modules/apps/feature-flag/feature-flag-web/src/main/resources/META-INF/resources/feature_flags.jsp
@@ -30,7 +30,7 @@ String displayStyle = featureFlagsDisplayContext.getDisplayStyle();
 				searchContainer="<%= featureFlagsDisplayContext.getSearchContainer() %>"
 			>
 				<liferay-ui:search-container-row
-					className="com.liferay.feature.flag.web.internal.model.FeatureFlagDisplay"
+					className="com.liferay.feature.flag.web.internal.display.FeatureFlagDisplay"
 					keyProperty="key"
 				>
