--- modules/apps/feature-flag/feature-flag-web/src/test/java/com/liferay/feature/flag/web/internal/constants/FeatureFlagConstantsTest.java
@@ -1,32 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.feature.flag.web.internal.constants;
-
-import com.liferay.portal.test.rule.LiferayUnitTestRule;
-
-import org.junit.Assert;
-import org.junit.ClassRule;
-import org.junit.Rule;
-import org.junit.Test;
-
-/**
- * @author Drew Brokke
- */
-public class FeatureFlagConstantsTest {
-
-	@ClassRule
-	@Rule
-	public static final LiferayUnitTestRule liferayUnitTestRule =
-		LiferayUnitTestRule.INSTANCE;
-
-	@Test
-	public void testGetKey() {
-		Assert.assertEquals("feature.flag", FeatureFlagConstants.getKey());
-		Assert.assertEquals(
-			"feature.flag.foo.bar", FeatureFlagConstants.getKey("foo", "bar"));
-	}
-
-}
